feat: splits configs to avoid import of backend config in frontend leaking secrets
